requirements.txt
groq
python-dotenv
pandas
PyPDF2
reportlab
sqlalchemy
